backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
redis>=5.0,<6.0
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0
gunicorn>=22.0,<24.0
Pillow>=10.0,<12.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
channels[daphne]>=4.1,<5.0
channels-redis>=4.2,<5.0

